decomp: Singer::Poll_ 0% -> 99.2%
Implement the per-frame vocal polling routine. Declares
VocalNoteList::PitchAt(float) const in VocalNote.h (declaration only,
genuine member function the function depends on). Corrects unk54 to
float in Singer.h.

+++ src/band3/game/Singer.h
@@ -165,7 +165,7 @@ public:
     float unk48;
     float mScreamEnergyThreshold; // 0x4c
     float unk50;
-    int unk54;
+    float unk54;
     int unk58;
     float mFrameMicPitch; // 0x5c
     float unk60;
